requirements.txt
altair==4.2.0
anyio==3.6.2
appnope @ file:///Users/builder/ci_310/appnope_1642500616005/work
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-timeout==4.0.2
attrs==22.2.0
Babel==2.11.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
bokeh==2.4.3
bs4==0.0.1
cachetools==5.2.1
cairocffi==1.4.0
CairoSVG==2.6.0
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_477u68wvzm/croot/certifi_1671487773341/work/certifi
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorcet==3.0.1
comm @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_0b9r9i3b7k/croot/comm_1671231125581/work
commonmark==0.9.1
contourpy==1.0.7
cryptography==39.0.0
cssselect2==0.7.0
cycler==0.11.0
debugpy @ file:///Users/builder/ci_310/debugpy_1642501698574/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml==0.7.1
echarts==0.0.0
entrypoints @ file:///opt/concourse/worker/volumes/live/5eb4850e-dcbc-41ad-5f22-922bac778f70/volume/entrypoints_1649926457041/work
et-xmlfile==1.1.0
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
Faker==16.6.0
fastjsonschema==2.16.2
favicon==0.7.0
fonttools==4.38.0
fqdn==1.5.1
gitdb==4.0.10
GithubContributions==0.3.0
GitPython==3.1.30
google-api-core==2.11.0
google-api-python-client==2.73.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.58.0
holoviews==1.15.4
htbuilder==0.6.1
httplib2==0.21.0
hvplot==0.8.2
idna==3.4
importlib-metadata==6.0.0
ipykernel @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_4dybncc18w/croot/ipykernel_1671488388285/work
ipython @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_444myk5dgj/croot/ipython_1670919323426/work
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jaraco.classes==3.2.3
jedi @ file:///opt/concourse/worker/volumes/live/18b71546-5bde-4add-72d1-7d16b76f0f7a/volume/jedi_1644315243726/work
Jinja2==3.1.2
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-bokeh==3.0.5
jupyter-events==0.6.3
jupyter_client @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_8dsejoebqs/croot/jupyter_client_1671703053245/work
jupyter_core @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_c1_m4s19su/croot/jupyter_core_1672332229812/work
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab==3.5.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
jupyterlab_server==2.18.0
keyring==23.13.1
keyrings.alt==4.2.0
kiwisolver==1.4.4
lxml==4.9.2
Markdown==3.4.1
markdownlit==0.0.5
MarkupSafe==2.1.1
matplotlib==3.6.3
matplotlib-inline @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_9ddl71oqte/croots/recipe/matplotlib-inline_1662014471815/work
mistune==2.0.4
more-itertools==9.0.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.3
nest-asyncio @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_64pfm74mxq/croot/nest-asyncio_1672387129786/work
node==1.2
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.1
oauthlib==3.2.2
odict==1.9.0
openpyxl==3.0.10
packaging==23.0
pandas==1.5.2
pandocfilters==1.5.0
panel==0.14.2
param==1.12.3
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
peppercorn==0.6
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.4.0
platformdirs @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_7fs8_2xgrm/croots/recipe/platformdirs_1662711383474/work
plotly==5.12.0
plumber==1.7
prettytable==3.6.0
prometheus-client==0.15.0
prompt-toolkit @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_82emz7mook/croot/prompt-toolkit_1672387300396/work
protobuf==3.20.3
psutil @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_c9b604bf-685f-47f6-8304-238e4e70557e1o7mmsot/croots/recipe/psutil_1656431274701/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyct==0.4.8
pydantic==1.10.4
pydeck==0.8.0
pyecharts==2.0.1
pygal==3.0.0
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pyicloud==1.0.0
pymdown-extensions==9.9.1
Pympler==1.0.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-decouple==3.7
python-google-calendar-api==0.5
python-json-logger==2.0.4
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pyviz-comms==2.2.1
PyYAML==6.0
pyzmq==25.0.0
redis==4.4.2
requests==2.28.2
requests-oauth==0.4.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.1.0
rsa==4.9
semver==2.13.0
Send2Trash==1.8.0
simplejson==3.18.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
spotipy==2.22.0
st-annotated-text==3.0.0
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
streamlit==1.17.0
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.4
streamlit-echarts==0.4.0
streamlit-embedcode==0.1.2
streamlit-extras==0.2.5
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.3
streamlit-keyup==0.2.0
streamlit-player==0.1.5
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
tenacity==8.1.0
terminado==0.17.1
tinycss2==1.2.1
todoist-api-python==2.0.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_1fimz6o0gc/croots/recipe/tornado_1662061695695/work
tqdm==4.64.1
traitlets @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_0dtilxc0bw/croot/traitlets_1671143889152/work
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.14
validators==0.20.0
watchdog==2.2.1
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
widgetsnbextension==4.0.5
withings-api==2.4.0
xyzservices==2022.9.0
zipp==3.11.0
zope.component==5.1.0
zope.deferredimport==4.4
zope.deprecation==4.4.0
zope.event==4.6
zope.hookable==5.4
zope.interface==5.5.2
zope.lifecycleevent==4.4
zope.proxy==4.6.1
